requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
Django==4.2.17
django-redis==5.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
redis==5.2.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
